+++ editor/editor-core/src/__tests__/visual-regression/common/quick-insert.ts
@@ -1,7 +1,10 @@
 import { snapshot, initEditorWithAdf, Appearance } from '../_utils';
 import { Page } from '../../__helpers/page-objects/_types';
 import { typeInEditorAtEndOfDocument } from '../../__helpers/page-objects/_editor';
-import { waitForMenuIconsToLoad } from '../../__helpers/page-objects/_quick-insert';
+import {
+  waitForTypeAheadMenu,
+  waitForMenuIconsToLoad,
+} from '../../__helpers/page-objects/_quick-insert';
 
 describe('Quick Insert:', () => {
   let page: Page;
@@ -15,7 +18,7 @@ describe('Quick Insert:', () => {
 
   it('should render the quick insert menu', async () => {
     await typeInEditorAtEndOfDocument(page, '/');
-    await page.waitForSelector('.fabric-editor-typeahead');
+    await waitForTypeAheadMenu(page);
     await waitForMenuIconsToLoad(page);
 
     await snapshot(page);