+++ editor/editor-core/src/__tests__/visual-regression/table/floating-toolbar.ts
@@ -4,17 +4,25 @@ import {
   clickFirstCell,
   clickTableOptions,
   clickCellOptions,
-  getSelectorForTableCell,
   selectCellOption,
 } from '../../__helpers/page-objects/_table';
+import { waitForFloatingControl } from '../../__helpers/page-objects/_toolbar';
+
+const floatingControlsAriaLabel = 'Table floating controls';
+const dropdownListSelector =
+  '[aria-label="Popup"] [data-role="droplistContent"]';
 
 describe('Table floating toolbar:fullpage', () => {
   let page: any;
 
+  // FIXME: toolbar centering isn't right...
   beforeEach(async () => {
     page = global.page;
     await initFullPageEditorWithAdf(page, adf);
+    // Focus the table and select the first (non header row) cell
     await clickFirstCell(page);
+    // Wait for floating table controls underneath the table
+    await waitForFloatingControl(page, floatingControlsAriaLabel);
   });
 
   afterEach(async () => {
@@ -22,16 +30,30 @@ describe('Table floating toolbar:fullpage', () => {
   });
 
   it('display options', async () => {
+    // Click table options within the floating toolbar
     await clickTableOptions(page);
+    // Wait for the drop down list within floating table controls to be shown
+    await page.waitForSelector(
+      `[aria-label="${floatingControlsAriaLabel}"] ${dropdownListSelector}`,
+    );
   });
 
   it('display cell options', async () => {
-    await getSelectorForTableCell({ row: 2, cell: 2 });
+    // Click table cell options button within the selected cell
     await clickCellOptions(page);
+    // Wait for table cell options drop down list to be shown
+    await page.waitForSelector(dropdownListSelector);
   });
 
   it('display cell background', async () => {
-    await getSelectorForTableCell({ row: 2, cell: 2 });
+    // Wait for table cell options drop down list to be shown, then
+    // select background color option and wait for color picker popout to be shown
     await selectCellOption(page, 'Cell background');
+    await page.waitForSelector(
+      `${dropdownListSelector} .pm-table-contextual-submenu`,
+    );
+    await page.waitForSelector(
+      `${dropdownListSelector} .pm-table-contextual-submenu`,
+    );
   });
 });